requirements.txt
streamlit
tensorflow
numpy
pillow
pandas
opencv-python-headless